requirements.txt
setuptools~=65.5.1
PyTerminal~=0.0.1
pyserial~=3.5